refactor: cardNumber 디렉터리 구조화
